% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimeWindows.R
\name{makeTimeWindows}
\alias{makeTimeWindows}
\title{makeTimeWindows}
\usage{
makeTimeWindows(start.date = "2000-01-01", end.date = today(),
  period = "m", market = "cn", offset = 0)
}
\arguments{
\item{offset}{}
}
\description{
makeTimeWindows
}
\examples{
makeTimeWindows("2016-01-01",today(),"q")
}
