% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimeWindowsHalfYearly.R
\name{makeTimeWindowsHalfYearly}
\alias{makeTimeWindowsHalfYearly}
\title{makeTimeWindowsHalfYearly}
\usage{
makeTimeWindowsHalfYearly(start.date = "2000-01-01", end.date = today(),
  market = "cn", offset = 0)
}
\arguments{
\item{market}{\code{market = c("cn", "hk")}}
}
\value{
\code{data.frame(start.date, end.date, period)}
}
\description{
make time window half-yealy
}
\examples{
makeTimeWindowsHalfYearly("2000-01-01", today())
}
