% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimeWindowsMonthly.R
\name{makeTimeWindowsMonthly}
\alias{makeTimeWindowsMonthly}
\title{makeTimeWindowsMonthly}
\usage{
makeTimeWindowsMonthly(start.date = "2017-01-01", end.date = Sys.Date(),
  market = "cn", offset = 0)
}
\arguments{
\item{market}{\code{market = c("cn", "hk")}}
}
\value{
\code{data.frame(start.date, end.date, period)}
}
\description{
make a monthly time window
start.date, end.date and period based on weekdays
previous.date based on tradable days
}
\examples{
makeTimeWindowsMonthly("2010-01-01",today())
}
