% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTimeWindowsWeekly.R
\name{makeTimeWindowsWeekly}
\alias{makeTimeWindowsWeekly}
\title{makeTimeWindowsWeekly}
\usage{
makeTimeWindowsWeekly(start.date = "2017-01-01", end.date = Sys.Date(),
  market = "cn")
}
\arguments{
\item{market}{\code{market = c("cn", "hk")}}
}
\value{
\code{data.frame(start.date, end.date, period)}
}
\description{
make a weekly time window
start.date, end.date and period based on weekdays
previous.date based on tradable days
}
\examples{
makeTimeWindowsWeekly("2010-01-01",today())
}
